just use petrinet-parts to reduce prop passing
