+++ packages/client/hmi-client/src/components/model/tera-model.vue
@@ -59,11 +59,16 @@
 				:model="temporaryModel"
 				@update-model="updateTemporaryModel"
 			/>
-			<tera-model-parts
-				class="mt-0"
-				:feature-config="featureConfig"
+			<tera-petrinet-parts
 				:model="temporaryModel"
-				@update-model="updateTemporaryModel"
+				:mmt="mmt"
+				:mmt-params="mmtParams"
+				:feature-config="featureConfig"
+				@update-state="(e: any) => onUpdate('state', e)"
+				@update-parameter="(e: any) => onUpdate('parameter', e)"
+				@update-observable="(e: any) => onUpdate('observable', e)"
+				@update-transition="(e: any) => onUpdate('transition', e)"
+				@update-time="(e: any) => onUpdate('time', e)"
 			/>
 		</section>
 	</tera-asset>
@@ -89,13 +94,22 @@ import TeraAsset from '@/components/asset/tera-asset.vue';
 import TeraAssetEnrichment from '@/components/widgets/tera-asset-enrichment.vue';
 import TeraInputText from '@/components/widgets/tera-input-text.vue';
 import TeraModelDescription from '@/components/model/petrinet/tera-model-description.vue';
-import TeraModelParts from '@/components/model/tera-model-parts.vue';
+import TeraPetrinetParts from '@/components/model/petrinet/tera-petrinet-parts.vue';
 import TeraSaveAssetModal from '@/components/project/tera-save-asset-modal.vue';
-import { getModel, updateModel } from '@/services/model';
+import { getModel, updateModel, getMMT } from '@/services/model';
 import type { FeatureConfig } from '@/types/common';
 import { AssetType, type Model } from '@/types/Types';
 import { useProjects } from '@/composables/project';
 import { logger } from '@/utils/logger';
+import { MiraModel, MiraTemplateParams } from '@/model-representation/mira/mira-common';
+import { emptyMiraModel } from '@/model-representation/mira/mira';
+import {
+	updateState,
+	updateParameter,
+	updateObservable,
+	updateTransition,
+	updateTime
+} from '@/model-representation/service';
 
 const props = defineProps({
 	assetId: {
@@ -118,8 +132,12 @@ const props = defineProps({
 
 const emit = defineEmits(['close-preview', 'on-save']);
 
+// const MMT_INJECTION_KEY = Symbol('mmt');
+
 const model = ref<Model | null>(null);
 const temporaryModel = ref<Model | null>(null);
+const mmt = ref<MiraModel>(emptyMiraModel());
+const mmtParams = ref<MiraTemplateParams>({});
 
 const newName = ref('New Model');
 const isRenaming = ref(false);
@@ -204,13 +222,57 @@ async function updateModelName() {
 	onSave();
 }
 
+async function updateMMT() {
+	if (!temporaryModel.value) return;
+	const response = await getMMT(temporaryModel.value);
+	if (!response) return;
+	mmt.value = response.mmt;
+	mmtParams.value = response.template_params;
+}
+
 function updateTemporaryModel(newModel: Model) {
+	let doMmtUpdate = false;
+	// Only update the MMT when the semantics of the model changes
+	if (
+		!isEqual(temporaryModel.value?.model, newModel.model) ||
+		!isEqual(temporaryModel.value?.semantics, newModel.semantics)
+	) {
+		doMmtUpdate = true;
+	}
 	temporaryModel.value = cloneDeep(newModel);
+	if (doMmtUpdate) updateMMT();
+}
+
+function onUpdate(property: string, event: any) {
+	if (!temporaryModel.value) return;
+	const newModel = cloneDeep(temporaryModel.value);
+	const { id, key, value } = event;
+	switch (property) {
+		case 'state':
+			updateState(newModel, id, key, value);
+			break;
+		case 'parameter':
+			updateParameter(newModel, id, key, value);
+			break;
+		case 'observable':
+			updateObservable(newModel, id, key, value);
+			break;
+		case 'transition':
+			updateTransition(newModel, id, key, value);
+			break;
+		case 'time':
+			updateTime(newModel, key, value);
+			break;
+		default:
+			break;
+	}
+	updateTemporaryModel(newModel);
 }
 
 async function fetchModel() {
 	model.value = await getModel(props.assetId);
 	temporaryModel.value = cloneDeep(model.value);
+	updateMMT();
 }
 
 onMounted(async () => {
